플레이어 정보 API를 위해 관련 Model 코드 작성.

+++ src/main/java/com/albionhelper/helper/domain/battle/EventPlayer.java
@@ -34,10 +34,10 @@ public class EventPlayer implements Comparator<EventPlayer> {
     private String allianceName;
 
     @JsonProperty("KillFame")
-    private int killFame;
+    private long killFame;
 
     @JsonProperty("DeathFame")
-    private int deathFame;
+    private long deathFame;
 
     @JsonProperty("DamageDone")
     private int damageDone;
